fixed variable type for vm-constant serial input #bugfix

+++ firmware/AVR-Source/Pyr0_Piezo_Sensor_v2.x.x/src/pP_serial.h
@@ -121,7 +121,7 @@ void updateHysteresis() {
 void updateConstant() {
   if (serialInt >= 0)
   {
-    voltMeterConstant = serialInt;
+    voltMeterConstant = (long) serialInt;
     EEPROM.put(VM_CONST_ADDRESS, voltMeterConstant);
   }
 }
